Rename zipformer2 to zipformer_for_ncnn_export_only to avoid confusion.
