+++ egs/csj/ASR/pruned_transducer_stateless7_streaming/zipformer_for_ncnn_export_only.py
@@ -1,0 +1,1 @@
+../../../librispeech/ASR/pruned_transducer_stateless7_streaming/zipformer_for_ncnn_export_only.py
