--- egs/librispeech/ASR/pruned_transducer_stateless7_streaming_multi/zipformer2.py
@@ -1,1 +1,0 @@
-../pruned_transducer_stateless7_streaming/zipformer2.py
